More nitpick in refreshDataSizeCrc: RRR
allSubs was all subpackages except u'' - so made that clear in the code.
Still uncertain what happens when I dump in top level "subpackage"
various types of stuff - readmes seem to be picked up - but that's under
266
The three elifs were not needed - if any of them true then
`elif not rootLower and reModExtMatch(fileExt)` would be false (so
they do not block entering that if) and the file will be skipped as
skip=True anyway (even entering the last elif). Take some uses of
the Installer skip variables for the ride - always welcome.