Adding extra cache with lines numbers - WIP:
Ugly (duplicates strings) but chops off another use of
getTweakFileSettings - based on the observation that whenever
it's called with lineNumbers=True is called on self.path ;)
Also the case with setCorrupted, which however now will be true in
ini.getTweakFileLines calls where (target) ini happens to be the tweak
(possible and I do not remove it - structures should be robust enough to
handle that ok). In this case setCorrupted will now be true - does not
harm.
We can improve memory performance on the cache any old time once stable,
note it's _only_ accessed inside getSettings.
TODO: OBSEIniFile edit that too