Skyrim - handle desync with plugins txt: RRR EEE algorithm Move ___savefixed here
Now Bash checks to see if the fetched loadorder.txt is not in sync with
cached_active (must be valid and reflect plugins.txt !). It then
proceeds to reorder the active mods present in lo in the same order as
in plugins.txt and saves the sync'ed loadorder.
For the algorithm see http://stackoverflow.com/q/30504317/281545 - that
question was motivated by this problem here)

Se also a discussion on which file should be sync'ed with what -
especially the posts by @ zilav:
http://forums.bethsoft.com/topic/1599816-wrye-bash-thread-108/page-3#entry25132038
and @ Arthmoor:
http://forums.bethsoft.com/topic/1599816-wrye-bash-thread-108/page-4#entry25132514
They both agree that plugins.txt should be given priority.

EDIT: needs testing for get (see _save_fixed_load_order commit)
and also the algorithm itself needs rethinking (see next commit) -
currently:

plugins = [ A, E, B ]

[ A, B, C, D, E, Z] -> [ A, E, C, D, B, Z ]

Better : [ A, C, D, E, B, Z ] (have the out of order "bubble up" so
they do not loose masters, if their master is an active moved mod
and they are active they should be reordered correctly if initially
ordered correctly - if inactive we must warn anyway - # )
(done later)

In getting for skyrim the load order must have already be synced.