_save_fixed_load_order needs not fetch the active plugins - WIP: REORDER OBLIVION TXT ONCE RRR commitSHA
See also    commit
on deleted files handling. If the client requests a refresh knowing (as
he does) that mods were deleted, he should flush his cache of active
mods - in other words pass force_active=True in GetLo. If the load order
is externally reordered however the fetching mechanism for Skyrim will
align it with the ones in the plugins.txt so this is taken care of.
Finally for fallout the operation is atomic anyway and will chop off
deleted anyway.

Mopy/bash/basher/mod_links.py - Espm flip:
We now have however to explicitly flush active cache cause if an active
mod was esp/m'fied we should (for skyrim) update plugins txt to account
for a possible change to this mod's order.
Took the opportunity to factor out common code in those links, no less
to factor out an ugly use of bosh.modInfos.plugins.refreshLoadOrder()
than to make this hack stand out - we should not be fetching but
setting load order here - we should however have _fix_load_order
exposed which seems not a good idea.