[!]Unused 'readMe' Installer persistent attribute -> 'extras_dict': RRR EEE
EEE rootIdex unused

Tricky I have to reason carefully - this is the only chance to hack in
more attributes to address design flaws (like the wizard images bug # )
and add functionality (like docs # ) so I have to get it right.
- 'extras_dict' sounds like a good name to me - extras is already there,
plus it is not really descriptive.
- Dropped `if self.dirty_sizeCrc is None` - under settings cruft (# ).
- reverting to older versions won't hurt as 'readMe' was unused - we
just set it, never used it.
- upgrading however needs hacking `__setstate__`

The rest is for fixing the bug reported here:

The rest is there to fix the bug brought to light by

whereupon on a layout like:

Project\
  some_folder\
    some.espm etc
  fomod\
    some_document.xml

the root index would be calculated as len(r'some_folder\') as fomod\
was skipped but then refreshDataSizeCrc would add file 'ocument.xml'
leading to file not found tracebacks and perpetual missing files (and
whatnot). I say "brought to light" and not "introduced", cause the same
would happen with files/folders starting with u'--' - just not as
common (I however remember some user having problems with
--packages).

Now _find_root_index updates extras_dict with the full path of the root,
will come handy in the wizards bug too - but I have to test all paths
self.fileSizeCrcs is used.

Signed-off-by: MrD <the.ubik@gmail.com>